Add files via upload
The Quiz Management System is a Java-based application designed to facilitate the creation, management, and evaluation of quizzes. It provides an intuitive and interactive interface for administrators, teachers, and students, making the process of quiz management more efficient and user-friendly. The system leverages Java's object-oriented features to ensure modularity and reusability, along with a back-end database for secure data storage and retrieval.